Update the SPI buffer to only send the two bytes of weight information, rather than pass along the full four bytes it gets from the floor.
Two of the four bytes will always be zero, so there's no need to transmit them.  Since the two bytes represent a
10-bit value, only 10 of the 16 bits in those two bytes are used, so we still have a valid stop condition (0xFF, 0xFF)

Finally, handle serial synchronously rather than running a thread.  Running a thread seemed to cause problems and made things more
complex to program and debug.  Its main goal was to make sure we weren't blocked on IO when we could be rendering a frame but that
doesn't seem to be an issue.