Added ability to start/stop the playlist
This is useful for setting up the floor since there are some
syncing issues around the data sent to the floor.  If the floor is powered on while data is
being sent it can easily get offset/shifted from the encoded data stream and decode starting
at the wrong place.  The result is that the lights freak out.  The fix is to stop the show, repower the floor
and then start the show again.