floor: processor: auto-register processors.
Removes the silly boilerplate we had gotten in to when
defining new processors. Also adds a unitest which ensures all
the processors in the default playlist are resolvable.

Currently assumes (a) all processors reside in the `processor`
submodule, and (b) all processor class names are globally unique;
which seem like safe enough assumptions for a while.