controller: static fps (2/2): clamp processors to desired fps.
The previous change boosted the call rate for `get_next_frame()` to
120/sec globally. But most existing processors were written to
expect the call rate they defined in `self.fps`, and so such
processors now render too fast.

This change restores that functionality but accomplishes it
differently, by extending the `@clocked()` decorator to accept
a `frames_per_second` clamp.