allow the controller to run and render even if playlist is stopped.
As a side-effect, this removes the discrete playlist start/stop api.

Instead, the playlist's RenderLayer enabled/disabled state governs
whether the playlist is on (running/advancing and rendering) or off (not
advancing, and not rendering anything).

This allows us to run the floor in modes where we only utilize the
overlay layers.