Cherry picking fix to head back to the flask ui
Add a timeout for the floor probe.

This is needed for when the floor is not actually attached to the Pi (testing etc), or not currently powered up.  In these cases the probe will hang forever waiting for packets to return.  In this case its preferable for it to give up after some time, in this commit, 2 seconds.

There's probably a better way to do this timeout (lib, etc) but opting for a zero dependency increase for this commit so that I can safely use it this weekend and not be debugging this small detail.

This has been tested on the Pi and the time out works.