Comitting the codebase as found on the production RaspberryPi
It was running an OS that was too old to load a recent version of git.  Most of the changes seem to be linting types of changes.  Going to start fresh, but keep this copy here just incase there was anything important and as a way to have something to go back to.