Clear the USI counter after chip select goes low
Very hopeful this will fix the instability issues with the tile.  When the floor gets in this state it seems like its not shifting data along the 4 byte boundary for each square. My theory is that the enable line goes high (disabled) before a full byte is read in to the USIDR register, leaving behind partial data and a starting point in the middle of the register when the next byte is sent.  ONce this happens, it will be off by X until the floor is power cycled.

By clearing the USI counter after select goes high, regardless of what we read in, it should guaruntee that we're in a good, known state to accept new data.