Tile stability improvements (#78)
* Clear the USI counter after chip select goes low

Very hopeful this will fix the instability issues with the tile.  When the floor gets in this state it seems like its not shifting data along the 4 byte boundary for each square. My theory is that the enable line goes high (disabled) before a full byte is read in to the USIDR register, leaving behind partial data and a starting point in the middle of the register when the next byte is sent.  ONce this happens, it will be off by X until the floor is power cycled.

By clearing the USI counter after select goes high, regardless of what we read in, it should guaruntee that we're in a good, known state to accept new data.

* PlatformIO doesn't support the ATTiny861 PLL clock at 16MHz.

Creating separate directories for platformio and for avrdude.  I'm moving back to avrdude for tile code, but keeping the platformio setup in case anything changes.

* Commit the last version of the platformio.ini file I used

* The VS Code stuff belongs with platformIO

* Changes made to two build files, one for avrdude and another for platformio

The platformio change actually got a build to be successful.

The avrdude change got it to recognize the programmer, but somehow it still couldn't recognize/program the AVR chip.

* Adding some platformio meta files now that it seems to be working

* Created an analyzer for Logic2 from Salaea

This analyzer understands the structure of the SPI data sent and prints out

* the square number
* the RGB values sent to each

* Added in a check for the chip select before continuing

Combined with clearing the USI counter, ideally this keeps us from getting into an inconsistent state

* Haha, omg, the dancefloor and RPi aboslutely cannot handle 120 FPS.  This backs up the Flask server, causing it to take > 5s to respond to the /api/status call which the FE calls every two seconds, which builds up a steady stream of open requests which eventually consumes all memory on the system.